Template
Template used for all the feature pages (phrasebook, weather, etc.);
this will keep consistency. Just make sure the menu stays in the bottom
part of the screen if you make your browser small enough for the phone,
and is in the top left corner if it’s at normal size.